+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [4.4.0] - 2022-08-16
 ## Changed
-- SwiftUI already scales fonts so scaling as part of the typography style View modifier e.g. `.typography(style: .interactive)` resulted in fonts being scaled twice. Therefore the default scaling mode when using this modifier in SwiftUI is disabled however the scaling mode can be specified as a parameter e.g. `.typography(style: .interactive, scalingMode: .fontMetrics)`.
+- When using the View modifier in SwiftUI, the scaling mode can be specified as a parameter e.g. `.typography(style: .interactive, scalingMode: .fontMetrics)`. By default the scaling mode specified in the configuration will be applied.
 
 ## [4.3.2] - 2022-03-16
 ## Changed